So this is broken at this point, but there's a bunch of working code in here. I broke it to fix it, but I need to save this.

+++ guinput/input_types.go
@@ -2,5 +2,5 @@ package guinput
 
 type KeyEvent struct {
   Kind  int
-  Key   int
+  Key   uint8
 }
